Implement background audio
This uses background media playback to let Discord run in the background during voice chat. It works by playing some sound through a MediaPlayer when a PeerConnection is active.

It also pauses Discord's AudioContext when the last PeerConnection is closed. This is needed because Windows thinks the app is still playing audio and keeps the app running in the background until every last audio stream is closed.