Shim MessageEvent.source for reCAPTCHA
Discord removes the reCAPTCHA widget without resetting it, causing reCAPTCHA to ignore messages from the iframe in Edge due to a failed `iframe.contentWindow == event.source` check and preventing the challenge from closing properly.

This makes MessageEvent.source return undefined if the iframe is removed, mimicking the behavior in Chrome.