Switch to Heroku, rewrite server in Go
This replaces the Google App Engine Python web proxy with a faster, portable version written in Go. It uses the standard net/http package and can be deployed to Heroku without leaving the web browser.

Because Go expects all code to be kept in a single workspace, the UWP projects are moved to a separate Git repo.